backend/requirements.txt
numpy
munkres
django
djangorestframework
django-cors-headers
Pillow
requests
redis
wechatpayv3